switched to data card encoding and json format
